+++ app/src/main/java/com/example/androidseminar/activity/SignUpActivity.kt
@@ -1,16 +1,14 @@
-package com.example.androidseminar
+package com.example.androidseminar.activity
 
 import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
 import com.example.androidseminar.data.RequestSignUpData
 import com.example.androidseminar.data.ResponseSignUpData
-import com.example.androidseminar.databinding.ActivitySignInBinding
 import com.example.androidseminar.databinding.ActivitySignUpBinding
 import com.example.androidseminar.util.BaseActivity
-import com.example.androidseminar.util.ServiceCreator
+import com.example.androidseminar.api.ServiceCreator
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
