+++ app/src/main/java/com/example/androidseminar/fragments/FollowerFragment.kt
@@ -2,7 +2,6 @@ package com.example.androidseminar.fragments
 
 import android.os.Bundle
 import android.util.Log
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -13,6 +12,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.example.androidseminar.R
 import com.example.androidseminar.adapter.FollowerRecyclerViewAdapter
+import com.example.androidseminar.api.GithubCreator
 import com.example.androidseminar.data.Info
 import com.example.androidseminar.data.ResponseGithubFollowerData
 import com.example.androidseminar.databinding.FragmentFollowerBinding
@@ -27,17 +27,7 @@ class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>(
     private lateinit var adapter: FollowerRecyclerViewAdapter
     private lateinit var itemTouchHelper: ItemTouchHelper
 
-    val followerData = mutableListOf(
-        Info("https://yt3.ggpht.com/ytc/AKedOLTBmVN3RYeIJpA6Rlmx1vloR3PGaDYR6sfoCTb4=s900-c-k-c0x00ffffff-no-rj", "윤현지", "안드YB"),
-    Info("https://mblogthumb-phinf.pstatic.net/MjAyMTAxMzFfMjIg/MDAxNjEyMDM1NDcyOTE4.3cRndWUdyeLevLJ3AmljQXVcIHbD_mX0hbra4PVkxgsg.0zn8FDRU_c28VWf8qhCO7Zy1v6QN6dnttfSGSLvrwssg.JPEG.letyourselfglow/IMG_1699.JPG?type=w800", "홍길동", "안드OB"),
-    Info("https://mblogthumb-phinf.pstatic.net/MjAyMDExMzBfMjY1/MDAxNjA2NzAxNzIzMDkz.mXpHE4L6KPe-c9o5QYP6pml_ck7Up-OsPeMR-cRemCUg.YKL7EaPVhrpTnTB2giNhEn6pVF6SvYb-LLRvp8ulNRcg.JPEG.marxela/IMG_0511.JPG?type=w800", "김길동", "서버YB"),
-    Info("https://i.pinimg.com/550x/c2/95/7d/c2957d833f37315e4dd4c51a34ce046c.jpg", "박길동", "아요YB")
-    )
-
-//    Info("https://yt3.ggpht.com/ytc/AKedOLTBmVN3RYeIJpA6Rlmx1vloR3PGaDYR6sfoCTb4=s900-c-k-c0x00ffffff-no-rj", "윤현지", "안드YB"),
-//    Info("https://mblogthumb-phinf.pstatic.net/MjAyMTAxMzFfMjIg/MDAxNjEyMDM1NDcyOTE4.3cRndWUdyeLevLJ3AmljQXVcIHbD_mX0hbra4PVkxgsg.0zn8FDRU_c28VWf8qhCO7Zy1v6QN6dnttfSGSLvrwssg.JPEG.letyourselfglow/IMG_1699.JPG?type=w800", "홍길동", "안드OB"),
-//    Info("https://mblogthumb-phinf.pstatic.net/MjAyMDExMzBfMjY1/MDAxNjA2NzAxNzIzMDkz.mXpHE4L6KPe-c9o5QYP6pml_ck7Up-OsPeMR-cRemCUg.YKL7EaPVhrpTnTB2giNhEn6pVF6SvYb-LLRvp8ulNRcg.JPEG.marxela/IMG_0511.JPG?type=w800", "김길동", "서버YB"),
-//    Info("https://i.pinimg.com/550x/c2/95/7d/c2957d833f37315e4dd4c51a34ce046c.jpg", "박길동", "아요YB")
+    val followerData = mutableListOf<Info>()
 
     override fun getFragmentBinding(
         inflater: LayoutInflater,
@@ -46,17 +36,15 @@ class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>(
         return FragmentFollowerBinding.inflate(inflater,container,false)
     }
 
-
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         initGithubFollowerNetwork()
         initFollowerRecyclerView()
-        //Log.d("followerData",followerData.toString())
+        Log.d("followerData",followerData.toString())
     }
 
     private fun initFollowerRecyclerView() {
         adapter = FollowerRecyclerViewAdapter(this)
-        adapter.infoList = followerData
-
+        //여기다가 adapter.infoList = followerData,adapter.notifyDataSetChanged() 해줬을땐 안됨!
         binding.followerRecyclerview.layoutManager = LinearLayoutManager(requireContext())
 
         itemTouchHelper = ItemTouchHelper(ItemTouchHelperCallback(adapter))
@@ -70,7 +58,6 @@ class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>(
             )
             )
         )
-
     }
 
     override fun onStartDrag(viewHolder: RecyclerView.ViewHolder) {
@@ -87,11 +74,12 @@ class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>(
             ) {
                 if(response.isSuccessful){
                     response.body()?.let {
-                        Log.d("Answers",it.toString())
-                        it.forEach { responseGithubFollowerData ->
-                            Log.d("Answers",responseGithubFollowerData.toString())
-                            //followerData.add(Info(responseGithubFollowerData.profileImg,responseGithubFollowerData.id))
+                        it.forEach { follower ->
+                            Log.d("Answers",follower.toString())
+                            followerData.add(Info(follower.profileImg,follower.id,"안드로이드"))
                         }
+                        adapter.infoList = followerData // TODO 왜 여기 위치해야함
+                        adapter.notifyDataSetChanged()
 
                     }
                 }else {
@@ -105,7 +93,4 @@ class FollowerFragment :ItemDragListener, BaseFragment<FragmentFollowerBinding>(
             }
         })
     }
-
-
-
 }
